Splitted poti and alsa into different files
